feat: moved validators and classes from controller
