+++ src/controllers/PenaltyDetailsController.ts
@@ -4,64 +4,9 @@ import { inject } from 'inversify';
 import { SessionService } from '../services/SessionService';
 import { CREATED, BAD_REQUEST, OK } from 'http-status-codes';
 import * as Joi from '@hapi/joi';
-
-const schema = Joi.object({
-    companyNumber: Joi.string()
-        .replace(' ', '')
-        .insensitive()
-        .uppercase()
-        .min(1).max(8)
-        .regex(/^(SC|NI)?[0-9]{6,8}/)
-        .messages({
-            'string.empty': 'You must enter a company number',
-            'string.min': 'You must enter your full eight character company number',
-            'string.max': 'You must enter your full eight character company number',
-            'string.pattern.base': 'You must enter your full eight character company number'
-        })
-    ,
-    penaltyReference: Joi.string()
-        .replace(' ', '')
-        .min(9).max(9)
-        .messages({
-            'string.empty': 'You must enter a penalty reference number',
-            'string.min': 'You must enter your reference number exactly as shown on your penalty notice',
-            'string.max': 'You must enter your reference number exactly as shown on your penalty notice'
-        })
-});
-
-const padNumber = (companyNumber: string): string => {
-    if (/^(SC|NI)/gm.test(companyNumber)) {
-        const leadingLetters = companyNumber.substring(0, 2);
-        let trailingChars = companyNumber.substring(2, companyNumber.length);
-        trailingChars = trailingChars.padStart(6, '0');
-        companyNumber = leadingLetters + trailingChars;
-    } else if (companyNumber.length > 0) {
-        companyNumber = companyNumber.padStart(8, '0');
-    }
-    return companyNumber;
-};
-
-
-
-class ValidationError {
-    constructor(public readonly field: string, public readonly text: string) { }
-
-    get href(): string {
-        return `#${this.field}-error`;
-    }
-
-}
-
-class ValidationResult {
-
-    constructor(public readonly errors: ValidationError[] = []) { }
-
-    public getErrorForField(field: string): ValidationError | undefined {
-
-        return this.errors.find(error => error.field == field);
-    }
-
-}
+import { schema, padNumber } from '../utils/PenaltyReferenceDetailsValidator';
+import { ValidationResult} from '../models/ValidationResult';
+import { ValidationError} from '../models/ValidationError';
 
 @controller('/penalty-reference')
 export class PenaltyDetailsController extends BaseHttpController {
